odmtp/requirements.txt
morph-kgc==2.8.0
Django==5.1
falcon==3.1.3
pymemcache==4.0.0
PyYAML==6.0.2
numpy==2.1.2
rdflib==7.1.1
python-dotenv==1.0.1